+++ defaults/controllers/_default/root/account/_index.js
@@ -1,0 +1,12 @@
+// --------------------
+// Account namespace controller
+// Route definition
+// --------------------
+
+// exports
+
+// route definition
+exports = module.exports = {
+	// route params
+	type: 'namespace'
+};