+++ defaults/controllers/_default/root/account/edit.js
@@ -1,0 +1,68 @@
+// --------------------
+// Account namespace controller
+// edit action
+// --------------------
+
+// libraries
+var forms = require('../../../../../lib/forms');
+
+// imports
+var defaultActionEdit = require('../../resource/edit');
+
+// exports
+
+// action definition
+exports = module.exports = {
+	// action params
+	actionTypes: {
+		form: true
+	},
+	
+	title: 'Change Account Details',
+	
+	// functions
+	
+	initForm: function() {
+		// make form
+		this.form = forms.createFormFromModel(this.route.overlook.models.user);
+		
+		// remove form fields
+		delete this.form.fields.type;
+		delete this.form.fields.isActive;
+	},
+	
+	load: function() {
+		return this.route.actions.index.load.call(this);
+	},
+	
+	populate: defaultActionEdit.populate,
+	
+	act: function() {
+		// init actResult
+		this.actResult = {};
+		
+		// record user id and date
+		this.actData.updatedById = this.user.id;
+		this.actData.updatedAt = new Date();
+		
+		// update db
+		return this.dataMain.updateAttributes(this.actData, {transaction: this.transaction}).bind(this)
+		.return(true)
+		.catch(this.sequelize.UniqueConstraintError, function(err) {
+			// unique field not unique
+			this.actResult = {error: 'uniqueFail', field: err.index};
+			return false;
+		})
+		.catch(this.sequelize.ForeignKeyConstraintError, function(err) {
+			// bad reference to another model
+			this.actResult = {error: 'illegalValue', field: err.index};
+			return false;
+		});
+	},
+	
+	done: function() {
+		return this.redirect('./', 'Saved changes to your account details');
+	},
+	
+	failed: defaultActionEdit.failed
+};